Vocabulary Extractor Update
The vocab extractor will now create 3 files: dothraki, english and a
dictionary for the two languages.
It also creates V2 and V3 variables for transitive verbs (vtr.) and V
and V2 variables for verbs marked as verb (v.) in the dothraki website
dictionary.
Spaces in variables will be replaced with underscores, as it is done in
the original resource grammar (be smelly -> be_smelly_V).