Added verbal auxiliaries can and must
"can" and "must are rendered as uninflected postpositions ("laz"
and "eth" respectively) to the subject of the sentence in Dothraki.

In GF these belong to the category VV (verbs with verb phrase
complement). However, apart from these two words (and a couple
more examples that are also rendered as particles), I could not
figure out how verbs with verb phrase complement usually work in
Dothraki (there are some indications that verb phrase complements
do not actually exist in Dothraki and have to be subsituted by
subordinate clauses, e.g. "The khaleesi wants to eat something
different" is translated as "Khaleesi zala meme adakha esinakh",
literally "The khaleesi wants that she eats something different").