Bugfix (?7 in DetQuant parses)
This was a weird bug, that I fixed by adding a new field, s, to
Num, which is always empty and adding the empty string from Num
to the linearization of Det in DetQuant. Why that would fix this
error, I have no idea (I found this fix through a lot of
experimentation with minimal examples).

Frankly, it seems like a bug in GF to me.